backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.38.0
sqlalchemy==2.0.48
alembic==1.18.4
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.3.3
xlsxwriter==3.2.0
beautifulsoup4==4.12.3
requests==2.32.2
python-dotenv==1.0.1
pydantic-settings==2.13.1
email-validator==2.2.0
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.7
